Separate https from cloud
